+++ manta-pay/src/crypto/hash/poseidon/parameter/mod.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with manta-rs.  If not, see <http://www.gnu.org/licenses/>.
 
-//! Poseidon hash parameter generation
+//! Poseidon Hash Parameter Generation
 
 pub mod lfsr;
 pub mod matrix;
